move table filters out to trait
