+++ app/Http/Livewire/Delegates/Delegates.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Http\Livewire\Delegates;
 
 use App\Http\Livewire\Concerns\DeferLoading;
+use App\Http\Livewire\Concerns\HasTableFilter;
 use App\Http\Livewire\Concerns\HasTablePagination;
 use App\Models\Scopes\OrderByBalanceScope;
 use App\Models\Wallet;
@@ -21,6 +22,7 @@ use Livewire\Component;
 final class Delegates extends Component
 {
     use DeferLoading;
+    use HasTableFilter;
     use HasTablePagination;
 
     const PER_PAGE = 51;
@@ -28,28 +30,10 @@ final class Delegates extends Component
     // TODO: Filters - https://app.clickup.com/t/861n4ydmh - see WalletTransactionTable
     public array $filter = [];
 
-    public bool $selectAllFilters = true;
-
     /** @var mixed */
     protected $listeners = [
         'setDelegatesReady' => 'setIsReady',
     ];
-
-    public function __get(mixed $property): mixed
-    {
-        if (array_key_exists($property, $this->filter)) {
-            return $this->filter[$property];
-        }
-
-        return parent::__get($property);
-    }
-
-    public function __set(string $property, mixed $value): void
-    {
-        if (array_key_exists($property, $this->filter)) {
-            $this->filter[$property] = $value;
-        }
-    }
 
     public function queryString(): array
     {
@@ -59,14 +43,6 @@ final class Delegates extends Component
 
     public function mount(bool $deferLoading = true): void
     {
-        foreach ($this->filter as &$filter) {
-            if (in_array($filter, ['1', 'true', true], true)) {
-                $filter = true;
-            } elseif (in_array($filter, ['0', 'false', false], true)) {
-                $filter = false;
-            }
-        }
-
         if (! $deferLoading) {
             $this->setIsReady();
         }
@@ -79,11 +55,6 @@ final class Delegates extends Component
         ]);
     }
 
-    public function getIsAllSelectedProperty(): bool
-    {
-        return ! collect($this->filter)->contains(false);
-    }
-
     // TODO: Filters - https://app.clickup.com/t/861n4ydmh - see WalletTransactionTable#getNoResultsMessageProperty
     public function getNoResultsMessageProperty(): null|string
     {
@@ -92,20 +63,6 @@ final class Delegates extends Component
         }
 
         return null;
-    }
-
-    public function updatedSelectAllFilters(bool $value): void
-    {
-        foreach ($this->filter as &$filter) {
-            $filter = $value;
-        }
-    }
-
-    public function updatedFilter(): void
-    {
-        $this->selectAllFilters = $this->isAllSelected;
-
-        $this->setPage(1);
     }
 
     public function getDelegatesProperty(): LengthAwarePaginator
@@ -122,7 +79,7 @@ final class Delegates extends Component
             ->paginate($this->perPage);
     }
 
-    // TODO: don't show if "active" filter is not selected - https://app.clickup.com/t/861n4ydmh
+    // TODO: return false if "active" filter is not selected - https://app.clickup.com/t/861n4ydmh
     public function getShowMissedBlocksProperty(): bool
     {
         if ($this->page > 1) {
